requirements.txt
numpy==1.26.2
pandas==2.1.4
py4j==0.10.9.7
pyspark==3.5.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
tzdata==2023.3
